+++ src/main/kotlin/racingcar/RacingCar.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.ui.InputView
+
 fun main() {
     val inputView = InputView()
     print(inputView.inputNumberOfCar())
