+++ src/main/kotlin/racingcar/ui/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.ui
 
 class InputView {
