+++ src/test/kotlin/racingcar/OutputViewTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import racingcar.ui.OutputView
 
 class OutputViewTest {
